% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation functions.R
\name{adj.to.pcor}
\alias{adj.to.pcor}
\title{find approximate pcor given an adjacency matrix and parameter range (again taken from bootnet: https://github.com/SachaEpskamp/bootnet/tree/master)}
\usage{
adj.to.pcor(adj, parRange = c(0.5, 1))
}
\arguments{
\item{adj}{the signed adjacency matrix}

\item{parRange}{the range of absolute partial correlations to start from (actual pcors will likely be lower than this)}
}
\value{
true.net = the partial correlation matrix
}
\description{
find approximate pcor given an adjacency matrix and parameter range (again taken from bootnet: https://github.com/SachaEpskamp/bootnet/tree/master)
}
