% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdit.network.R
\name{cdit.network}
\alias{cdit.network}
\title{Estimate a dependency network using conditional distance independence tests with multivariate control}
\usage{
cdit.network(data, boots, chores)
}
\arguments{
\item{data}{data as matrix or data frame}

\item{boots}{number of bootstraps for local bootstrapping P value computation}
}
\value{
adjacency matrix
}
\description{
This function takes data and estimates the dependence structure using conditional
distance independence tests. Missing data is removed.
}
