% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation functions.R
\name{compare.sim.networks}
\alias{compare.sim.networks}
\title{Compare simulated etworks}
\usage{
compare.sim.networks(true, est, metric = NULL, directed = FALSE)
}
\arguments{
\item{true}{true network}

\item{est}{estimated network}

\item{metric}{which performance measures to use}

\item{directed}{is the network directed or not}
}
\description{
This function computes various performance measures to be used in simulation studies
}
\details{
this is a function from https://osf.io/preprints/psyarxiv/4j3hf_v1 added to this package for convenience to be used
in simulation studies
}
