% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcor.network.R
\name{dcor.network}
\alias{dcor.network}
\title{Estimate a zero order dependency network from distance correlations}
\usage{
dcor.network(
  data,
  select = "sig",
  dcor.permutations = NULL,
  alpha = 0.05,
  adjust = "fdr",
  adjust.threshold = 0.1
)
}
\arguments{
\item{data}{data as matrix or data frame}

\item{select}{method for present/absent edge selection. Can either be "saturated" for no thresholding, "sig" for significance testing (default), of "qp" for q1-graph structure estimation}

\item{dcor.permutations}{if "sig" for select, how many permutations should be performed for the permutation test for each edge weight (default = 1000).}

\item{alpha}{if "sig" for select, the alpha level for thresholding the edges (default = 0.05)}

\item{adjust}{which method should be used for P value adjustment? default FDR using BH procedure}

\item{adjust.threshold}{what threshold should be used for the adjusted P value? default is 10\% FDR}
}
\value{
saturated.net = The saturated dcor network

selected.net = The thresholded network if select = "sig" (or saturated network if select = "saturated")

p.vals = p value matrix if thresholding is selected

edge.set = edge list with estimates and P values
}
\description{
This function takes data and estimates the zero order distance correlation network
It returns the saturated dcor network (used if network moderation analysis performed)
as well as the 'selected' network by thresholding at FDR < 0.1 (user can change this setting).
Missing data is handled using pairwise complete observations.
}
