% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{degree}
\alias{degree}
\title{Node degree}
\usage{
degree(net)
}
\arguments{
\item{net}{The estimated network structure}
}
\value{
Degree of each node. Only relevant in thresholded networks.
}
\description{
Node degree
}
