% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.dpm.R
\name{estimate.dpm}
\alias{estimate.dpm}
\title{Estimate a dependency network with distance precision network}
\usage{
estimate.dpm(data, select = "k-means")
}
\arguments{
\item{data}{data as matrix or data frame}

\item{select}{method for present/absent edge selection. Can either be "saturated" for no thresholding, "k-means" for clustering, of "qp" for q1-graph structure estimation}

\item{threshold}{if qp graph is method for edge selection, what is the threshold of the non-rejection rate to select the graph structure (default = 0.95)? The alpha level argument is used for individual q tests}
}
\value{
saturated.net = The saturated pdcor network

selected.net = The thresholded network (or saturated network if select = "saturated")

edge.set = edge list with estimates
}
\description{
This function takes data and estimates the distance precision network. Missing data is removed.
}
