% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.kendall.R
\name{estimate.kendall}
\alias{estimate.kendall}
\title{Estimate a kendall based dependency network}
\usage{
estimate.kendall(data, select = "sig", alpha = 0.05, threshold = 0.95)
}
\arguments{
\item{data}{data as matrix or data frame}

\item{select}{method for present/absent edge selection. Can either be "saturated" for no thresholding, "sig" for significance testing (default), of "qp" for q1-graph structure estimation}

\item{alpha}{if "sig" for select, the alpha level for thresholding the edges (default = 0.05)}

\item{threshold}{if qp graph procedure selected, threshold for NRR}
}
\value{
saturated.net = The saturated pdcor network

selected.net = The thresholded network if select = "sig" (or saturated network if select = "saturated")

p.vals = p value matrix if thresholding is selected

edge.set or edge.signs = edge list with estimates and P values
}
\description{
Estimate a kendall based dependency network
}
