% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{extract.property}
\alias{extract.property}
\title{Extract perms}
\usage{
extract.property(test, property)
}
\arguments{
\item{property}{the network statistic to extract}

\item{perms}{the permuted data object}
}
\value{
the permuted null distribution for the request network statistic
}
\description{
Extract perms
}
