% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation functions.R
\name{genGGM}
\alias{genGGM}
\title{Generate GGM function. From (cite). Used for simulation studies.}
\usage{
genGGM(
  Nvar,
  p = 0,
  nei = 1,
  parRange = c(0.5, 1),
  constant = 1.5,
  propPositive = 0.5,
  clusters = NULL,
  graph = c("smallworld", "random", "scalefree", "hub", "cluster")
)
}
\arguments{
\item{Nvar}{number of nodes}

\item{p}{rewiring probability}

\item{nei}{average neighbourhood of each node}

\item{parRange}{range of partial correlations}

\item{constant}{default = 1.5}

\item{propPositive}{Proportion of positive partial correlations}

\item{clusters}{number of clusters ig graph = "cluster"}

\item{graph}{one of smallworld, random, scalefree, hub, or cluster}
}
\value{
partial correlation matrix
}
\description{
Generate GGM function. From (cite). Used for simulation studies.
}
