% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{global.cc}
\alias{global.cc}
\title{Average clustering coefficient}
\usage{
global.cc(net)
}
\arguments{
\item{net}{The estimated network structure}
}
\description{
Average clustering coefficient
}
