% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{global.weighted.cc}
\alias{global.weighted.cc}
\title{Average weighted clustering coefficient}
\usage{
global.weighted.cc(net)
}
\arguments{
\item{net}{The estimated network structure}
}
\value{
Average weighted clustering coefficient
}
\description{
Average weighted clustering coefficient
}
