% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{hamming.distance}
\alias{hamming.distance}
\title{Hamming distance}
\usage{
hamming.distance(net1, net2)
}
\arguments{
\item{net1}{The estimated network structure}

\item{net2}{The estimated network structure for group 2}
}
\value{
Estimated Hamming dissimilarity. Only relevant for thresholded unweighted networks
}
\description{
Hamming distance
}
