% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{jaccard.similarity}
\alias{jaccard.similarity}
\title{Jaccard dissimilarity}
\usage{
jaccard.similarity(net1, net2)
}
\arguments{
\item{net1}{The estimated network structure for group 1}

\item{net2}{The estimated network structure for group 2}
}
\value{
Estimated Jaccard dissimilarity. Only relevant for thresholded unweighted networks
}
\description{
Jaccard dissimilarity
}
