% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.boot.R
\name{net.boot}
\alias{net.boot}
\title{Perform bootstrap resampling of network estimate}
\usage{
net.boot(data, net, nBoots, stability = TRUE)
}
\arguments{
\item{data}{the dataset used to estimate the emperical network (as matrix or data frame)}

\item{net}{the estimate.network object which contains the saturated and selected network estimates}

\item{nBoots}{the number of bootstrap resamples to perform. Default is 1000.}

\item{stability}{Should a network structure be selected on each resample such that edge weight stability can be calculated?}
}
\value{
stability results = a dataframe with parameters for each edge weight

boot.object = the object from the boot package
}
\description{
This function performs bootstrapping analysis (using the boot package) to assess the stability and accuracy of edge weights.
}
