% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.comparison.analysis.R
\name{network.comparison.analysis}
\alias{network.comparison.analysis}
\title{Perform network comparison analysis of zero-order networks with a permutation test}
\usage{
network.comparison.analysis(data1, data2, net1, net2, permutations)
}
\arguments{
\item{data1}{the dataset as matrix or data frame for group 1}

\item{data2}{the dataset as matrix or data frame for group 2}

\item{net1}{the estimate.dcor.network object of group 1}

\item{net2}{the estimated.dcor.network object of group 2}

\item{permutations}{the number of permutations to perform. Default is 1000.}
}
\value{
dissimilarity.results = network.dissimilarity estimates, difference, P and S values

global.results = global network topology estimates, difference, P and S values

strength.results = Node strength estimates, difference, P and S values

edge.differences = Edge weights, difference, P and S values

perms = A list of all permuted null distributions for each network statistic
}
\description{
This function performs network comparison analysis. That is, it performs a permutation test to compare two networks
on various global, nodal, and edge level characteristics. It uses zero-order distance correlations.
}
\details{
Note that in the returned objects/results, both the P and S value are supplied. For dissimilarity
the P value is one-sided, for all other network descriptors it is two-sided. The S value is also
known as the surprisal value and is simply calculated as -log2(P value). It converts the P value
into binary 'bits' of information against the test hypothesis and can be interpreted in the context
of flipping a coin. That is, a P value of 0.05 as an S value is 4.32, rounded to the nearest integer
we can say that assuming the null hypothesis (and all other assumptions) are true, then the result
is as surprising as getting 4 heads in a row in a series of coin tosses. See Rafi and Greenland (2020)
for more information and explanation about the surprisal value (https://doi.org/10.1186/s12874-020-01105-9).
}
