% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.intervention.analysis.R
\name{network.intervention.analysis}
\alias{network.intervention.analysis}
\title{Estimate a q1 graph network for network intervention analysis}
\usage{
network.intervention.analysis(
  data,
  alpha = 0.1,
  nrr = FALSE,
  nrr.threshold = NULL,
  num.boots = 1000,
  chores = 1
)
}
\arguments{
\item{data}{data as matrix or data frame}

\item{alpha}{the alpha level for each individual test (default = 0.1)}

\item{nrr}{if this is equal to true then the non rejection rate procedure is used to select the network structure. Defaults to FALSE.}

\item{nrr.threshold}{if nrr is equal to TRUE this is the non rejection rate threshold for edge selection}

\item{num.boots}{number of bootstraps to calculate P value (default = 1000)}

\item{chores}{how many computer chores should be used when performing cdcor significance test}
}
\value{
structure = the undirected unweighted network structure. This is the network intervention analysis q1-graph network.

nrr = non-rejection rate matrix if nrr is equal to TRUE

alpha = the alpha level used in all significance tests
}
\description{
This function takes data which includes a binary intervention variable and estimates a q1-graph network
}
