% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics functions.R
\name{node.size}
\alias{node.size}
\title{This contains various functions for graphical displays}
\usage{
node.size(data, centrality, p1, p2)
}
\arguments{
\item{data}{The data used in the network analyis (extracts columns names from this)}

\item{centrality}{A vector with a centrality score for each node}

\item{p1}{value of parameter 1}

\item{p2}{value of parameter 2}
}
\value{
vector of node sizes to be supplied to qgraph
}
\description{
Change the node sizes given a node level descriptors
}
