% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn.transform.R
\name{npn.transform}
\alias{npn.transform}
\title{Transform non-gaussian continuous data to gaussian via the non-paranormal transformation}
\usage{
npn.transform(data)
}
\arguments{
\item{data}{Path to the input file}
}
\value{
a data frame of the same dataset after npn transformation
}
\description{
This function takes the raw non-gaussian data and applies the non-paranormal transformation.
Missing data is maintained in the same entries as the raw data. Only supply the variables
you want to transform.
}
