% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{one.sided.p.function}
\alias{one.sided.p.function}
\title{One sided P value from permuted null distribution}
\usage{
one.sided.p.function(perms, emp.diff)
}
\arguments{
\item{perms}{the permutation distribution}

\item{emp.diff}{the emperical estimate of the statistic}
}
\value{
the permutatin test one sided P value
}
\description{
One sided P value from permuted null distribution
}
