% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdcor.network.R
\name{pdcor.network}
\alias{pdcor.network}
\title{Estimate a partial distance correlation network}
\usage{
pdcor.network(
  data,
  select = "sig",
  dcor.permutations = NULL,
  alpha = 0.05,
  adjust = "none",
  adjust.threshold = NULL,
  qp.threshold = 0.95,
  zero.order = FALSE
)
}
\arguments{
\item{data}{data as matrix or data frame}

\item{select}{method for present/absent edge selection. Can either be "saturated" for no thresholding, "sig" for significance testing (default), of "qp" for q1-graph structure estimation}

\item{dcor.permutations}{if "sig" for select, how many permutations should be performed for the permutation test for each edge weight (default = 1000).}

\item{alpha}{if "sig" for select, the alpha level for thresholding the edges (default = 0.05)}

\item{adjust}{which method should be used for P value adjustment? default FDR using BH procedure}

\item{adjust.threshold}{what threshold should be used for the adjusted P value? default is 10\% FDR}

\item{qp.threshold}{if qp graph is method for edge selection, what is the threshold of the non-rejection rate to select the graph structure (default = 0.95)? The alpha level argument is used for individual q tests}
}
\value{
saturated.net = The saturated pdcor network

selected.net = The thresholded network if select = "sig" (or saturated network if select = "saturated")

p.vals = p value matrix if thresholding is selected

edge.set or edge.signs = edge list with estimates and P values

nrr = non-rejection rate matrix if "qp" is method of edge selection
}
\description{
This function takes data and estimates a network based on partial distance correlation
}
