% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{proportion.neg}
\alias{proportion.neg}
\title{Proportion negative edges. Only relevant for signed networks.}
\usage{
proportion.neg(net)
}
\arguments{
\item{net}{The estimated network structure}
}
\value{
Proportion of negative edges
}
\description{
Proportion negative edges. Only relevant for signed networks.
}
