% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{proportion.pos}
\alias{proportion.pos}
\title{Proportion positive edges. Only relevant for signed networks.}
\usage{
proportion.pos(net)
}
\arguments{
\item{net}{The estimated network structure}
}
\value{
Proportion of positive edges
}
\description{
Proportion positive edges. Only relevant for signed networks.
}
