% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp.mgm.R
\name{qp.mgm}
\alias{qp.mgm}
\title{Estimate a q1 graph network using node wise generalized linear models
#'}
\usage{
qp.mgm(data, alpha = 0.1, type, nrr = FALSE, nrr.threshold = NULL)
}
\arguments{
\item{data}{data as matrix or data frame}

\item{alpha}{the alpha level of the maximum P value to select edges}

\item{type}{character string of variable types e.g. "g" for gaussian, "b" for binary, "pr" for proportion/beta distributions, and "p" for Poisson}

\item{nrr}{if this is equal to true then the non rejection rate procedure is used to select the network structure. Defaults to FALSE and the maximum P value method is used.}

\item{nrr.threshold}{if nrr is equal to TRUE this is the non rejection rate threshold for edge selection}
}
\value{
structure = the undirected unweighted network structure

nrr = non-rejection rate matrix if nrr is equal to TRUE

alpha = the alpha level used for the maximum P value
}
\description{
Estimate a q1 graph network using node wise generalized linear models
#'
}
