% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy.detection.R
\name{redundancy.detection}
\alias{redundancy.detection}
\title{Identify possibly redundant nodes based on zero-order correlation}
\usage{
redundancy.detection(
  data,
  retain.nodes = NULL,
  retain.edges = NULL,
  threshold = 0.4,
  permutations = 1000
)
}
\arguments{
\item{data}{the data to estimate correlations from}
}
\value{
redundancy.results = dataframe with all dcors and suggested redundancies

sig.redundancy.results = dataframe with only pairs over threshold shown

dcor.net = pairwise zero-order distance correlation network/matrix

threshold = the threshold used to decide about possible redundancies
}
\description{
This function estimates pairwise zero-order distance correlations.
It returns the correlation matrix, a matrix of all pairwise results, and a matrix of possibly redundant
variables/nodes based on a user supplied correlation threshold. Users can also supply a character vector
containing the names of variables/nodes which should not be suggested to be redundant with any other node.
}
