% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{sF}
\alias{sF}
\title{Frobenius norm dissimilarity}
\usage{
sF(net1, net2, p)
}
\arguments{
\item{net1}{The estimated network structure}

\item{net2}{The estimated network structure for group 2}

\item{p}{The number of nodes}
}
\value{
Estimated Frobenius norm dissimilarity. Function adapted from:
}
\description{
Frobenius norm dissimilarity
}
