% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation functions.R
\name{sim.data}
\alias{sim.data}
\title{simulate data from a given GGM (again taken from bootnet: https://github.com/SachaEpskamp/bootnet/tree/master)}
\usage{
sim.data(trueNet, sampleSize)
}
\arguments{
\item{trueNet}{partial correlation matrix}

\item{sampleSize}{sample size}
}
\value{
data as matrix
}
\description{
simulate data from a given GGM (again taken from bootnet: https://github.com/SachaEpskamp/bootnet/tree/master)
}
