% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{strength}
\alias{strength}
\title{Node strength}
\usage{
strength(net)
}
\arguments{
\item{net}{The estimated network structure}
}
\value{
Strength of each node. Only relevant in weighted networks.
}
\description{
Node strength
}
