% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{two.sided.p.function}
\alias{two.sided.p.function}
\title{Two sided P value from permuted null distribution}
\usage{
two.sided.p.function(perms, emp.diff)
}
\arguments{
\item{perms}{the permutation distribution}

\item{emp.diff}{the emperical estimate of the statistic}
}
\value{
the permutatin test two sided P value
}
\description{
Two sided P value from permuted null distribution
}
