% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{unweighted.density}
\alias{unweighted.density}
\title{This contains various functions to compute different topological descriptors of networks.
They are called in the compare.networks function.}
\usage{
unweighted.density(net)
}
\arguments{
\item{net}{The estimated network structure}
}
\value{
Network density estimate
}
\description{
Global network density
}
