% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.descriptors.R
\name{weighted.density}
\alias{weighted.density}
\title{Global weighted density}
\usage{
weighted.density(net)
}
\arguments{
\item{net}{The estimated network structure}
}
\value{
Network weighted density estimate
}
\description{
Global weighted density
}
