% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.network.R
\name{estimate.network}
\alias{estimate.network}
\title{Estimate a network structure}
\usage{
estimate.network(estimator, ...)
}
\arguments{
\item{estimator}{= the network estimation algorithm to use}

\item{...}{= the arguments specific for the estimation function selected}
}
\value{
structure = the selected network structure
}
\description{
This function is a wrapper around the various network estimation procedures implemented in this package
and other network analysis R packages. See details for estimation algorithms implemented.
}
